% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNumeric.R
\name{analyzeNumeric}
\alias{analyzeNumeric}
\title{analyzeNumeric}
\usage{
analyzeNumeric(data, var, bivariate = FALSE)
}
\description{
\code{analyzeNumeric} Performs univariate analysis of numeric variables
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{back}}, \code{\link{bivariate}},
  \code{\link{forward}}, \code{\link{preprocess}},
  \code{\link{process}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
