% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateModel.R
\name{evaluateModel}
\alias{evaluateModel}
\title{evaluateModel}
\usage{
evaluateModel(movies, omit, predictors)
}
\arguments{
\item{movies}{Data frame with movie features}

\item{omit}{Numeric indices for variables to omit from the model}

\item{predictors}{Numeric indices for predictor columns}
}
\description{
\code{evaluateModel} evaluates a model given a response variable and
predictors
}
\seealso{
Other xmar functions: \code{\link{mlr}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
