% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{forward}
\usage{
forward(movies)
}
\description{
\code{forward} Creates prediction model using forward selection
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{analyzeNumeric}}, \code{\link{back}},
  \code{\link{learnAudienceSize}},
  \code{\link{preprocess}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
