% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnAudienceSize.R
\name{learnAudienceSize}
\alias{learnAudienceSize}
\title{learnAudienceSize}
\usage{
learnAudienceSize(movies)
}
\description{
\code{learnAudienceSize} Performs learnAudienceSizeing of data for analysis
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{analyzeNumeric}}, \code{\link{back}},
  \code{\link{forward}}, \code{\link{preprocess}},
  \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
