% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{mlr}
\usage{
mlr(movies)
}
\description{
\code{mlr} mlres the data set and creates data sets used
for analysis.
}
\seealso{
Other xmar functions: \code{\link{evaluateModel}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
