% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateQuant.R
\name{multivariateQuant}
\alias{multivariateQuant}
\title{multivariateQuant}
\usage{
multivariateQuant(data, yLab)
}
\arguments{
\item{data}{Single column data frame containing the quantitative variable}

\item{yLab}{Capitalized character string for the variable name or label}
}
\value{
analysis List containing:
 1. Summary Statistics
 2. Histogram
 3. Normal quantile plot
 4. Box plot
}
\description{
\code{multivariateQuant} Performs multivariate analysis on a single quantitative
variable.  The function returns descriptive statistics, a histogram
and a box plot to check outliers.
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
