% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCooks}
\alias{plotCooks}
\title{plotCooks}
\usage{
plotCooks(mod, yLab)
}
\arguments{
\item{mod}{Linear model object}

\item{yLab}{Capital case character string for response variable}
}
\value{
cooks Cooks distance plot for residuals
}
\description{
\code{plotCooks} Plots cooks distance
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotFreqProp}}, \code{\link{plotHist}},
  \code{\link{plotQQ}}, \code{\link{plotResAll}},
  \code{\link{plotResFit}}, \code{\link{plotResLeverage}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
