% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotResAll}
\alias{plotResAll}
\title{plotResAll}
\usage{
plotResAll(mod)
}
\arguments{
\item{mod}{Linear model object}
}
\value{
list containing ressidual v predictor plots
}
\description{
\code{plotResAll} Plots cooks distance
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotCooks}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotQQ}},
  \code{\link{plotResFit}}, \code{\link{plotResLeverage}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
