% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotResQQ}
\alias{plotResQQ}
\title{plotResQQ}
\usage{
plotResQQ(mod, yLab)
}
\arguments{
\item{mod}{Linear model object}

\item{yLab}{Capital case character string for response variable}
}
\value{
resQQ Residual Normal QQ Plot
}
\description{
\code{plotResQQ} Plots regression diagnostics
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotCooks}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotQQ}},
  \code{\link{plotResAll}}, \code{\link{plotResFit}},
  \code{\link{plotResLeverage}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
