% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionAnalysis.R
\name{regressionAnalysis}
\alias{regressionAnalysis}
\title{regressionAnalysis}
\usage{
regressionAnalysis(mod, xLab, yLab)
}
\arguments{
\item{mod}{Linear model}

\item{xLab}{Capitalized character string for the categorical variable}

\item{yLab}{Capitalized character string for the quantitative variable}
}
\value{
analysis plots, summary statistics, test results and interpretive text
}
\description{
\code{regressionAnalysis} Performs regression analysis
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
