% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(movies)
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{analyzeNumeric}}, \code{\link{back}},
  \code{\link{forward}}, \code{\link{learnAudienceSize}},
  \code{\link{preprocess}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
