% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeCategorical.R
\name{analyzeCategorical}
\alias{analyzeCategorical}
\title{analyzeCategorical}
\usage{
analyzeCategorical(data, var)
}
\description{
\code{analyzeCategorical} Performs univariate analysis of categorical variables
}
\seealso{
Other movies functions: \code{\link{analyzeNumeric}},
  \code{\link{back}}, \code{\link{forward}},
  \code{\link{learnAudienceSize}},
  \code{\link{preprocess}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
