% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNumeric.R
\name{analyzeNumeric}
\alias{analyzeNumeric}
\title{analyzeNumeric}
\usage{
analyzeNumeric(data, var)
}
\description{
\code{analyzeNumeric} Performs univariate analysis of numeric variables
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{back}}, \code{\link{forward}},
  \code{\link{learnAudienceSize}},
  \code{\link{preprocess}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
