% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back.R
\name{back}
\alias{back}
\title{back}
\usage{
back(movies)
}
\description{
\code{back} Creates prediction model using backward elimination
}
\seealso{
Other movies functions: \code{\link{bivariate}},
  \code{\link{forward}}, \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
