% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{bivariate}
\alias{bivariate}
\title{bivariate}
\usage{
bivariate(mdb)
}
\description{
\code{bivariate} Performs bivariate analysis of predictors on the response
variable.
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
