% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{bivariate}
\alias{bivariate}
\title{bivariate}
\usage{
bivariate(data)
}
\arguments{
\item{data}{List containing two data frames, mdb, the main movie data set and mdbBox the sample with box office revenue}
}
\description{
\code{bivariate} Performs bivariate analysis of variables
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
