% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFreqProp}
\alias{plotFreqProp}
\title{plotFreqProp}
\usage{
plotFreqProp(data, yLab = "Movies", xLab, title = NULL, order = NULL)
}
\arguments{
\item{data}{Data frame or vector containing a single categorical factor variable}

\item{yLab}{Capital case character string describing the y variable}

\item{xLab}{Capital case character string containing the name of the variable x variable}

\item{title}{Capital case character string for the title of the plot}

\item{order}{Character c('a', 'd') indicating whether to order by frequency ascending (a), or descending (d)}
}
\value{
List containing a contingency table and a stacked bar plot.
}
\description{
\code{plotFreqPropTbl} Renders a stacked bar plot and a contingency table
showing frequencies and proportions
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotCooks}}, \code{\link{plotHist}},
  \code{\link{plotQQ}}, \code{\link{plotResAll}},
  \code{\link{plotResFit}}, \code{\link{plotResLeverage}},
  \code{\link{plotResQQ}}, \code{\link{plotScatter}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
