% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotScatter}
\alias{plotScatter}
\title{plotScatter}
\usage{
plotScatter(data, xLab, yLab, title = NULL)
}
\arguments{
\item{data}{Data frame containing the quantitative variables}

\item{xLab}{Capital case character string containing the name of the grouping or subset variable}

\item{yLab}{Capital case character string containing the name of the y variable}

\item{title}{Capital case character string for title of plot}
}
\value{
Scatterplot object
}
\description{
\code{plotScatter} Renders a scatterplot for two numerical variablesa
}
\seealso{
Other visualization functions: \code{\link{plotBox}},
  \code{\link{plotCooks}}, \code{\link{plotFreqProp}},
  \code{\link{plotHist}}, \code{\link{plotQQ}},
  \code{\link{plotResAll}}, \code{\link{plotResFit}},
  \code{\link{plotResLeverage}}, \code{\link{plotResQQ}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
