% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(movies)
}
\description{
\code{preprocess} Performs preprocessing of data for analysis
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{analyzeNumeric}}, \code{\link{back}},
  \code{\link{forward}}, \code{\link{learnAudienceSize}},
  \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
