% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(mdb)
}
\description{
\code{process} Pepares the data for the modeling stage
}
\seealso{
Other movies functions: \code{\link{analyzeCategorical}},
  \code{\link{analyzeNumeric}}, \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{preprocess}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
