% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R, R/process.R
\name{process}
\alias{process}
\alias{process}
\title{process}
\usage{
process(mdb)

process(mdb)
}
\description{
\code{process} Pepares the data for the bivariate analysis stage

\code{process} Pepares the data for the bivariate analysis stage
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}

Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{univariateQual}},
  \code{\link{univariateQuant}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
