% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(mdb1, mdb2)
}
\arguments{
\item{mdb1}{Data frame containing the preprocessed movie data set}

\item{mdb2}{Data frame containing 100 samples from the mdb data frame with box office revenue added}
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{regressionAnalysis}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
