% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(mdb)
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other movies functions: \code{\link{back}},
  \code{\link{bivariate}}, \code{\link{forward}},
  \code{\link{multivariateQuant}},
  \code{\link{preprocess}}, \code{\link{process}},
  \code{\link{univariateQual}},
  \code{\link{univariateQuant}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
